Refactor orc bloom fitler test and both iceberg and hive can test
